+++ runtime/cudaq/algorithms/evolve.h
@@ -12,9 +12,9 @@
 #include "common/KernelWrapper.h"
 #include "cudaq/algorithms/get_state.h"
 #include "cudaq/base_integrator.h"
-#include "cudaq/dynamics/operator_type.h"
 #include "cudaq/host_config.h"
 #include "cudaq/operators.h"
+#include "cudaq/operators/operator_type.h"
 #include "cudaq/platform.h"
 #include "cudaq/platform/QuantumExecutionQueue.h"
 #include "cudaq/schedule.h"
