+++ runtime/cudaq/operators/helpers.cpp
