+++ runtime/cudaq/operators/helpers.h
