+++ runtime/cudaq/operators/operator_type.h
