+++ runtime/cudaq/operators/scalar_op.cpp
