+++ runtime/cudaq/operators/sum_op.cpp
